requirements.txt
# ANC Platform - Enterprise Backend Requirements
# Comprehensive dependencies for production deployment

# ============================================================================
# Core Audio Processing
# ============================================================================
pyaudio>=0.2.11
numpy>=1.25.2,<2.6
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# ============================================================================
# Machine Learning
# ============================================================================
scikit-learn>=1.3.0
joblib>=1.3.0

# ============================================================================
# Web Framework & API
# ============================================================================
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.5
Werkzeug>=3.0.0
gunicorn>=21.2.0
gevent>=23.9.1
gevent-websocket>=0.10.1

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# ============================================================================
# Database
# ============================================================================
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9  # PostgreSQL support

# ============================================================================
# Caching & Message Queue
# ============================================================================
redis>=5.0.1
celery>=5.3.4
flower>=2.0.1  # Celery monitoring

# ============================================================================
# HTTP & Networking
# ============================================================================
requests>=2.31.0
urllib3>=2.1.0
python-socketio>=5.10.0

# ============================================================================
# Configuration & Utilities
# ============================================================================
pyyaml>=6.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# ============================================================================
# Monitoring & Logging
# ============================================================================
prometheus-client>=0.19.0
prometheus-flask-exporter>=0.23.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-flask>=1.3.0
pytest-mock>=3.12.0
faker>=22.0.0

# ============================================================================
# Code Quality
# ============================================================================
black>=23.12.1
flake8>=7.0.0
pylint>=3.0.3
mypy>=1.8.0

# ============================================================================
# Optional: Visualization
# ============================================================================
matplotlib>=3.7.0

# ============================================================================
# Production Server
# ============================================================================
uvicorn[standard]>=0.25.0  # Alternative ASGI server
eventlet>=0.34.2  # Alternative async library
